WIP: Implement sharding down to the data layer
This commit makes sharding work all the way down to the database level.
It brings changes to the core database code, and a refactor of how
connections are gotten from the shared pool.

Testing framework has changed; we now run a single migration on a
template database, and then create a copy for every test we run. The
results are very fast integration tests, without worrying about
re-running the entire sharded migration every time, and without special
case code to change sharding behavior at test time.

The sessionsrv is migrated to PostgreSQL, and sharded on Account ID.

* Fix the other services to also use the new shard logic
* Move some functionality from originsrv into sessionsrv

Signed-off-by: Adam Jacob <adam@chef.io>